🚚 api: move certificationBadgesService to src/shared
