+++ api/src/certification/enrolment/domain/usecases/index.js
@@ -7,6 +7,7 @@ import * as organizationRepository from '../../../../shared/infrastructure/repos
 import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
 import { importNamedExportsFromDirectory } from '../../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
 import * as attendanceSheetPdfUtils from '../../../enrolment/infrastructure/utils/pdf/attendance-sheet-pdf.js';
+import * as certificationBadgesService from '../../../shared/domain/services/certification-badges-service.js';
 import * as temporaryCompanionStorageService from '../../../shared/domain/services/temporary-companion-storage-service.js';
 import { enrolmentRepositories } from '../../infrastructure/repositories/index.js';
 import * as certificationCpfService from '../services/certification-cpf-service.js';
@@ -14,7 +15,6 @@ import * as sessionCodeService from '../services/session-code-service.js';
 import * as sessionsImportValidationService from '../services/sessions-import-validation-service.js';
 import * as temporarySessionsStorageForMassImportService from '../services/temporary-sessions-storage-for-mass-import-service.js';
 import * as sessionValidator from '../validators/session-validator.js';
-import * as certificationBadgesService from '../../../../../lib/domain/services/certification-badges-service.js';
 
 /**
  * @typedef {import('../../infrastructure/repositories/index.js').ComplementaryCertificationRepository} ComplementaryCertificationRepository
@@ -36,6 +36,7 @@ import * as certificationBadgesService from '../../../../../lib/domain/services/
  * @typedef {import('../../infrastructure/utils/pdf/attendance-sheet-pdf.js')} AttendanceSheetPdfUtils
  * @typedef {import('../services/temporary-sessions-storage-for-mass-import-service.js').TemporarySessionsStorageForMassImportService} TemporarySessionsStorageForMassImportService
  * @typedef {import('../../../shared/domain/services/temporary-companion-storage-service.js')} TemporaryCompanionStorageService
+ * @typedef {import('../../../shared/domain/services/certification-badges-service.js')} CertificationBadgesService
  **/
 
 /**
@@ -62,6 +63,7 @@ import * as certificationBadgesService from '../../../../../lib/domain/services/
  * @typedef {SessionValidator} SessionValidator
  * @typedef {AttendanceSheetPdfUtils} AttendanceSheetPdfUtils
  * @typedef {TemporaryCompanionStorageService} TemporaryCompanionStorageService
+ * @typedef {CertificationBadgesService} CertificationBadgesService
  **/
 const dependencies = {
   certificationBadgesService,
