+++ api/tests/certification/shared/unit/domain/services/certification-badges-service_test.js
@@ -1,5 +1,5 @@
-import * as certificationBadgesService from '../../../../lib/domain/services/certification-badges-service.js';
-import { domainBuilder, expect, sinon } from '../../../test-helper.js';
+import * as certificationBadgesService from '../../../../../../src/certification/shared/domain/services/certification-badges-service.js';
+import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | Service | certification-badges-service', function () {
   describe('#findStillValidBadgeAcquisitions', function () {
